# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KigaliWeShare API — neighbor-to-neighbor carpool matching for Kigali commuters
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}/id-verified:
    post:
      operationId: setUserIdVerified
      tags: [admin]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetIdVerifiedBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [userId, idVerified]
                properties:
                  userId: { type: string }
                  idVerified: { type: boolean }
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, verified, all]
            default: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUserRow" }
  /admin/users/{userId}/verify:
    post:
      operationId: verifyUser
      tags: [admin]
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyUserBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserRow" }
  /admin/invites:
    get:
      operationId: listInvites
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InviteCode" }
    post:
      operationId: createInvite
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInviteBody" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InviteCode" }
  /admin/neighborhoods:
    post:
      operationId: createAdminNeighborhood
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateNeighborhoodBody" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Neighborhood" }
  /admin/neighborhoods/{id}:
    put:
      operationId: updateAdminNeighborhood
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateNeighborhoodBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Neighborhood" }
    delete:
      operationId: deleteAdminNeighborhood
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /admin/corridors:
    post:
      operationId: createAdminCorridor
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCorridorBody" }
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Corridor" }
  /admin/corridors/{id}:
    put:
      operationId: updateAdminCorridor
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCorridorBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Corridor" }
    delete:
      operationId: deleteAdminCorridor
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Deleted }
  /admin/config:
    put:
      operationId: updateConfig
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateConfigBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicConfig" }
components:
  schemas:
...
    Role:
      type: string
      enum: [rider, driver, both]
    AccountStatus:
      type: string
      enum: [pending, verified, suspended]
...
    Neighborhood:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        sector: { type: string, nullable: true }
    Corridor:
      type: object
      required: [id, label, originId, destinationId, distanceKm]
      properties:
        id: { type: string }
        label: { type: string, description: "e.g. Remera \u2013 CBD" }
        originId: { type: string }
        destinationId: { type: string }
        distanceKm: { type: number }
    PublicConfig:
      type: object
      required: [fuelPriceRwfPerLitre, dieselPriceRwfPerLitre, vehicleConsumptionLPer100Km, currencyCode]
      properties:
        fuelPriceRwfPerLitre: { type: number }
        dieselPriceRwfPerLitre: { type: number }
        vehicleConsumptionLPer100Km: { type: number }
        currencyCode: { type: string, description: "e.g. RWF" }
        rnpEmergencyNumber: { type: string }
        vapidPublicKey: { type: string }
...
    Language:
      type: string
      enum: [en, fr, rw]
...
    AdminUserRow:
      type: object
      required: [userId, fullName, role, status, createdAt]
      properties:
        userId: { type: string }
        fullName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/AccountStatus" }
        neighborhoodName: { type: string, nullable: true }
        employer: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        inviteLabel: { type: string, nullable: true }
        completedTrips: { type: integer }
        averageRating: { type: number, nullable: true }
        idVerified: { type: boolean }
        preferredLanguage: { $ref: "#/components/schemas/Language" }
        createdAt: { type: string, format: date-time }
    VerifyUserBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/AccountStatus" }
    InviteCode:
      type: object
      required: [id, code, label, maxUses, uses, createdAt]
      properties:
        id: { type: string }
        code: { type: string }
        label: { type: string }
        maxUses: { type: integer }
        uses: { type: integer }
        createdAt: { type: string, format: date-time }
    CreateInviteBody:
      type: object
      required: [label, count]
      properties:
        label: { type: string, description: "Employer/community tag, e.g. Norrsken Pilot" }
        count: { type: integer, minimum: 1, maximum: 100 }
        maxUses: { type: integer, minimum: 1, default: 1 }
    CreateNeighborhoodBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        sector: { type: string, nullable: true }
    UpdateNeighborhoodBody:
      type: object
      properties:
        name: { type: string }
        sector: { type: string, nullable: true }
    CreateCorridorBody:
      type: object
      required: [label, originId, destinationId, distanceKm]
      properties:
        label: { type: string }
        originId: { type: string }
        destinationId: { type: string }
        distanceKm: { type: number }
    UpdateCorridorBody:
      type: object
      properties:
        label: { type: string }
        originId: { type: string }
        destinationId: { type: string }
        distanceKm: { type: number }
    UpdateConfigBody:
      type: object
      properties:
        fuelPriceRwfPerLitre: { type: number }
        dieselPriceRwfPerLitre: { type: number }
        vehicleConsumptionLPer100Km: { type: number }
...
    SetIdVerifiedBody:
      type: object
      required: [idVerified]
      properties:
        idVerified: { type: boolean }
...