# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KigaliWeShare API — neighbor-to-neighbor carpool matching for Kigali commuters
 * OpenAPI spec version: 0.1.0
 */

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  rider: "rider",
  driver: "driver",
  both: "both",
} as const;